pyproject.toml
[project]
name = "disco-engine"
dynamic = ["version"]
description = "A distributed event simulation core (disco) engine."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [{ name = "Michiel Jansen" }]
classifiers = [
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: C++",
  "Operating System :: OS Independent",
]
dependencies = [
  "numpy~=2.3",
  "pandas~=2.3",
  "pydantic~=2.12",
  "pydantic-settings~=2.12",
  "kazoo~=2.10",
  "python-graphblas~=2025.2",
  "SQLAlchemy~=2.0",
  "disco-tools>=0.1.2",
  "disco-data-logger>=0.1.2",
  "grpcio>=1.60",
  "protobuf>=4.25",
  "PyYAML>=6.0"
]

[build-system]
requires = [
  "scikit-build-core>=0.11.6",
  "pybind11>=2.11",
  "numpy>=2.0",
  "setuptools-scm>=8",
]
build-backend = "scikit_build_core.build"

[tool.scikit-build]
editable.mode = "redirect"

[tool.scikit-build.wheel.packages]
disco = "src/disco"

[tool.scikit-build.metadata]
version.provider = "scikit_build_core.metadata.setuptools_scm"

[tool.setuptools_scm]
write_to = "src/disco/_version.py"
version_scheme = "guess-next-dev"
local_scheme = "no-local-version"
fallback_version = "0.0.0"

[project.optional-dependencies]
dev = [
  "pytest>=8",
  "pytest-cov>=5",
  "mypy>=1.10",
  "cibuildwheel>=2.20",
  "pybind11-stubgen>=2.5",
  "twine>=5",
  "grpcio-tools>=1.60",
]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-q"
testpaths = ["tests"]

[tool.cibuildwheel]
build = "cp311-* cp312-*"
skip = "*-win32 *-musllinux_*"
test-command = 'pytest -q "{project}/tests"'
test-extras = ["dev"]
environment = { PIP_ONLY_BINARY = "numpy", PIP_PREFER_BINARY = "1", PIP_DISABLE_PIP_VERSION_CHECK = "1" }

[tool.cibuildwheel.linux]
archs = ["x86_64"]

[tool.cibuildwheel.macos]
archs = ["arm64", "x86_64"]

[tool.scikit-build.cmake.define]
PYBIND11_FINDPYTHON = true
CMAKE_DOWNLOAD_TIMEOUT = 30
BUILD_EXT = true

[tool.mypy]
python_version = "3.12"
warn_unused_configs = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = true
ignore_missing_imports = true
check_untyped_defs = true
no_implicit_optional = true
strict_equality = true
show_error_codes = true

disallow_untyped_defs = false
disallow_incomplete_defs = false

[[tool.mypy.overrides]]
module = "disco.*"
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_reexport = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
